refactor auth
